add error status and invitations

+++ src/services/target.service.ts
@@ -2,6 +2,7 @@ import AppDataSource from "../config/db";
 import { CreateTargetDTO } from "../DTO/Target/createTarget.dto";
 import { UpdateTargetDTO } from "../DTO/Target/updateTarget.dto";
 import { Target } from "../entities/Target";
+import { AppError } from "../utils/AppError";
 
 const targetRepository = AppDataSource.getRepository(Target);
 
@@ -10,7 +11,7 @@ export class TargetService {
     const existing = await targetRepository.findOneBy({ name: target.name });
 
     if (existing) {
-      throw new Error(`L'audience avec le nom "${target.name}" existe déjà.`);
+      throw new AppError(`L'audience avec le nom "${target.name}" existe déjà.`, 409);
     }
     return await targetRepository.save(targetRepository.create(target));
   }
@@ -18,7 +19,7 @@ export class TargetService {
   async updateTarget(targetId: number, dto: UpdateTargetDTO) {
     const targetToUpdate = await targetRepository.findOneBy({ id: targetId });
     if (!targetToUpdate) {
-      throw new Error(`Audience avec l'ID ${targetId} introuvable.`);
+      throw new AppError(`Audience avec l'ID ${targetId} introuvable.`, 404);
     }
     if (dto.name !== undefined) {
       targetToUpdate.name = dto.name;
@@ -32,10 +33,18 @@ export class TargetService {
   }
 
   async findTargetById(targetId: number) {
-    return await targetRepository.findOneBy({ id: targetId });
+    const target = await targetRepository.findOneBy({ id: targetId })
+    if (!target) {
+    throw new AppError("Audience non trouvée", 404);
+  }
+    return target
   }
 
   async deleteTarget(targetId: number) {
+    const target = await targetRepository.findOneBy({ id: targetId })
+    if (!target) {
+      throw new AppError("Audience non trouvée", 404);
+    }
     const res = await targetRepository.delete(targetId);
     return res.affected !== 0;
   }
